% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_metric.R
\name{dataset_metric}
\alias{dataset_metric}
\title{Dataset Metric}
\usage{
dataset_metric(data, privileged_groups, unprivileged_groups)
}
\arguments{
\item{data}{A StructuredDataset.}

\item{privileged_groups}{Privileged groups. List containing privileged protected attribute name and value of the privileged protected attribute.}

\item{unprivileged_groups}{Unprivileged groups. List containing unprivileged protected attribute name and value of the unprivileged protected attribute.}
}
\description{
Class for computing metrics based on one StructuredDataset.
}
\examples{
\dontrun{
load_aif360_lib()
# Input dataset
data <- data.frame("feat" = c(0,0,1,1,1,1,0,1,1,0), "label" = c(1,0,0.5,1,0,0,0.5,0,0.5,1))
# Create structured dataset
sd <- aif360::structured_dataset(data_path = data, target_column="label",
                                 protected_attribute="feat",
                                 unprivileged_protected_attribute=0,
                                 privileged_protected_attribute=1
)
# Create an instance of dataset metric
dm <- dataset_metric(sd, list('feat', 1), list('feat',2))
# Access metric functions
dm$num_instances()
}
}
\seealso{
\href{https://aif360.readthedocs.io/en/latest/modules/metrics.html#dataset-metric}{Explore available dataset metrics here}

Available metrics:
\itemize{
  \item difference
  \item num_instances
  \item ratio
}
}
