% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_aif360}
\alias{install_aif360}
\title{Install aif360 and its dependencies}
\usage{
install_aif360(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  version = "default",
  envname = NULL,
  extra_packages = NULL,
  restart_session = TRUE,
  conda_python_version = "3.7",
  ...
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows. Note also
that since this command runs without privilege the "system" method is
available only on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{version}{AIF360 version to install. Specify "default" to install
the latest release.}

\item{envname}{Name of Python environment to install within}

\item{extra_packages}{Additional Python packages to install.}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{conda_python_version}{the python version installed in the created conda
environment. Python 3.6 is installed by default.}

\item{...}{other arguments passed to [reticulate::conda_install()] or
[reticulate::virtualenv_install()].}
}
\description{
Install aif360 and its dependencies
}
