% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_datasets.R
\name{law_school_gpa_dataset}
\alias{law_school_gpa_dataset}
\title{Law School GPA Dataset}
\usage{
law_school_gpa_dataset()
}
\description{
Law School GPA Dataset
}
\seealso{
\href{https://aif360.readthedocs.io/en/latest/modules/generated/aif360.datasets.LawSchoolGPADataset.html#aif360.datasets.LawSchoolGPADataset}{More about the Law School GPA dataset.}
}
