% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inprocessing_prejudice_remover.R
\name{prejudice_remover}
\alias{prejudice_remover}
\title{Prejudice Remover}
\usage{
prejudice_remover(eta=1.0, sensitive_attr='',class_attr='')
}
\arguments{
\item{eta}{fairness penalty parameter}

\item{sensitive_attr}{name of protected attribute}

\item{class_attr}{label name}
}
\description{
Prejudice remover is an in-processing technique that adds a discrimination-aware regularization term to the learning objective
}
\examples{
\dontrun{
# An example using the Adult Dataset
load_aif360_lib()
ad <- adult_dataset()
model <- prejudice_remover(class_attr = "income-per-year", sensitive_attr = "race")
model$fit(ad)
ad_pred <- model$predict(ad)
}
}
