% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_dataset.R
\name{structured_dataset}
\alias{structured_dataset}
\title{AIF360 Structured Dataset}
\usage{
structured_dataset(data_path, target_column, protected_attribute,
                     unprivileged_protected_attribute,
                     privileged_protected_attribute)
}
\arguments{
\item{data_path}{Path to the input CSV file or a R dataframe.}

\item{target_column}{Name describing the label.}

\item{protected_attribute}{A feature for which fairness is desired.}

\item{unprivileged_protected_attribute}{An unprotected attribute value which is considered privileged from a fairness perspective.}

\item{privileged_protected_attribute}{A protected attribute value which is considered privileged from a fairness perspective.}
}
\description{
Function to create AIF structured dataset.
}
\examples{
\dontrun{
load_aif360_lib()
# Input dataset
data <- data.frame("feat" = c(0,0,1,1,1,1,0,1,1,0), "label" = c(1,0,0.5,1,0,0,0.5,0,0.5,1))
# Create aif structured dataset
act <- aif360::structured_dataset(data_path = data, target_column="label",
                                  protected_attribute="feat",
                                  unprivileged_protected_attribute=0,
                                  privileged_protected_attribute=1
                                 )
}
}
\seealso{
\href{https://aif360.readthedocs.io/en/latest/modules/generated/aif360.datasets.StructuredDataset.html#aif360.datasets.StructuredDataset}{More about AIF structured dataset.}
}
